refactor(mcp): move MCP server to mcp_server/ at project root
Rename evidencelab_mcp/ → mcp_server/ to:
- Place MCP server alongside ui/ and pipeline/ as a top-level module
- Avoid name collision with the mcp pip package (no sys.path hacks)
- Rename server.py → app.py to avoid shadowing mcp.server

All imports, docker-compose command, and tests updated.

Tested: MCP container starts, tools/list returns 3 tools, search
returns results. All 15 unit tests pass.
