+++ mcp_server/auth.py
