+++ mcp_server/tools/assistant.py
@@ -6,7 +6,7 @@ import asyncio
 import logging
 from typing import Any, Dict, List, Optional
 
-from evidencelab_mcp.schemas import MCPAssistantResponse
+from mcp_server.schemas import MCPAssistantResponse
 
 logger = logging.getLogger(__name__)
